backend/requirements.txt
# fastapi==0.104.1
# uvicorn==0.24.0
# python-multipart==0.0.6
# numpy==1.24.3
# pandas==2.0.3
# pillow==10.0.1
# scikit-learn==1.3.0
# dvc==3.48.4
# azure-monitor-query==1.2.0
# azure-identity==1.15.0
# pytest==7.4.3
# # Use PyTorch versions compatible with Python 3.13 (CPU)
# torch==2.5.1
# torchvision==0.20.1
# torchaudio==2.5.1
# opencv-python-headless==4.8.1.78
# Core API packages
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Data & preprocessing
numpy==1.24.3
pandas==2.0.3
pillow==10.0.1
scikit-learn==1.3.0
opencv-python-headless==4.8.1.78

# Lightweight, CPU-only PyTorch stack
torch==2.5.1+cpu
torchvision==0.20.1+cpu
torchaudio==2.5.1+cpu
--extra-index-url https://download.pytorch.org/whl/cpu

# Optional dev / Azure tools (keep only if you need them)
dvc==3.48.4
azure-monitor-query==1.2.0
azure-identity==1.15.0
pytest==7.4.3
